requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-Admin==2.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.28
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
WTForms==3.2.1
